:recycle: Refactor location of types into their own file and update the related imports in other classes
