+++ src/types/types.ts
@@ -1,0 +1,18 @@
+export type BaseStats = {
+  hp: number;
+  attack: number;
+  defence: number;
+  specialDefence: number;
+  specialAttack: number;
+  speed: number;
+};
+
+export type Pokemon = {
+  name: string;
+  id: string;
+  sprite: string;
+  height: number;
+  weight: number;
+  types: string[];
+  baseStats: BaseStats;
+};
